requirements.txt
alembic==1.4.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-CLI==0.4.0
Flask-DotEnv==0.1.2
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
pycodestyle==2.5.0
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
requests==2.22.0
six==1.14.0
SQLAlchemy==1.3.13
stripe==2.42.0
urllib3==1.25.8
Werkzeug==0.16.1
WTForms==2.2.1
